Ensure that TextDecoder.decode supports an ArrayBuffer if provided.
The current implementation although hinted as supporting an ArrayBuffer
isn't actually dealing with one when provided - since 1.0.2 the
default handling is that the value is an ArrayBufferView - the calls
afterwards leverage buffer.buffer (and similar) and thus the decode
fails.

A few new conditionals have been added which also provide a null value
as a fallback position so that methods can always be called.

Additionally the basic test has been altered slightly too so that we
use the existing Uint8Array and retrieve its ArrayBuffer to verify
that too can be passed through the decode method.